+++ src/main/java/ctrlS/totori/global/security/JwtTokenProvider.java
@@ -1,8 +1,6 @@
-package ctrlS.totori.global.jwt;
+package ctrlS.totori.global.security;
 
-import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
 import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.security.Keys;
 import org.springframework.beans.factory.annotation.Value;
