+++ src/main/java/ctrlS/totori/relation/repository/ParentChildRepository.java
@@ -1,5 +1,6 @@
-package ctrlS.totori.domain.parentchild;
+package ctrlS.totori.relation.repository;
 
+import ctrlS.totori.relation.entity.ParentChild;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ParentChildRepository extends JpaRepository<ParentChild, Long> {
